+++ controllers/front/ajax.php
@@ -96,7 +96,7 @@ class ItellashippingAjaxModuleFrontController extends ModuleFrontController
     $carrier = new Carrier($order->id_carrier);
 
     if (!$this->isItellaOrder($carrier)) {
-      echo "Not Itella order.";
+      echo "Not Smartpost order.";
       die;
     }
 
@@ -262,8 +262,12 @@ class ItellashippingAjaxModuleFrontController extends ModuleFrontController
       $itellaCart->updateItellaCartTrackNumber($data['id_cart'], $tracking_number);
       // save tracking number(s) to order carrier as well
       $order->setWsShippingNumber($tracking_number);
+      $order->shipping_number = $tracking_number;
+      $order->update();
 
-      return json_encode(array('success' => 'Itella API: Order registered.', 'filename' => $id_order . '.pdf', 'tracking_number' => $tracking_number));
+      ItellaShipping::changeOrderStatus($id_order, ItellaShipping::getCustomOrderState());
+
+      return json_encode(array('success' => 'Smartpost API: Order registered.', 'filename' => $id_order . '.pdf', 'tracking_number' => $tracking_number));
     } catch (ItellaException $e) {
       $itellaCart->saveError($data['id_cart'], $e->getMessage());
       return json_encode(array('errors' => $e->getMessage()));